Begin refactoring type system before continuing AST refactor. Things will be broken for a little bit. Use previous commits to enjoy the language until resolved.

+++ tests/sha256_test.fx
@@ -244,7 +244,7 @@ def main()->int
     print();
     print(ctx.state[1]);
     print();
-    print((u32)ctx.state[2]);
+    print(ctx.state[2]);
     print();
     print(ctx.state[3]);
     print();
